% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{gene.EM.velocity.estimates}
\alias{gene.EM.velocity.estimates}
\title{Estimate Region velocity using EM algorithm}
\usage{
gene.EM.velocity.estimates(
  vel,
  iteration = 100,
  deltaT = 1,
  n.cores = defaultNCores(),
  verbose = TRUE
)
}
\arguments{
\item{vel}{- velocity estimation(from gene.relative.velocity.estimates)}

\item{iteration}{- Max steps of EM iterations}

\item{deltaT}{- amount of time to project the cell forward}

\item{n.cores}{- number of cores to use}

\item{verbose}{- output messages about progress}
}
\value{
a list with velocity results, including the current normalized expression state ($current), projected ($projected) over a certain time ($deltaT), unscaled transcriptional change ($deltaE), fit results ($gamma, $ko, $sfit if spanning reads were used), optional cell pooling parameters ($cellKNN, $kCells), kNN-convolved normalized matrices (conv.imat.norm and conv.emat.norm), library scale ($mult)
}
\description{
Estimate Region velocity using EM algorithm
}
\examples{
\dontrun{
 # use EM algorithm to infer parameters of analytical solution. 
}
}
