% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lv_alpha_color}
\alias{lv_alpha_color}
\title{Adjust colour transparency, keeping vector names}
\usage{
lv_alpha_color(x, alpha = 1, ...)
}
\arguments{
\item{x}{colour vector}

\item{alpha}{transparency passed to \code{adjustcolor} as \code{alpha.f}}

\item{...}{further arguments to \code{adjustcolor}}
}
\value{
colour vector with original names
}
\description{
Adjust colour transparency, keeping vector names
}
\examples{
lv_alpha_color(c("red", "blue"), alpha = 0.5)
}
