% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lv_col_cor}
\alias{lv_col_cor}
\title{Pairwise correlations of matrix columns}
\usage{
lv_col_cor(m)
}
\arguments{
\item{m}{dense numeric matrix}
}
\value{
correlation matrix
}
\description{
Pairwise correlations of matrix columns
}
\examples{
m <- matrix(rnorm(30), ncol = 3)
lv_col_cor(m)
}
