% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp.R
\name{lv_em_objective_one_sided}
\alias{lv_em_objective_one_sided}
\title{One-sided EM objective (Eq 18)}
\arguments{
\item{t}{hidden cell times}

\item{Sobs, Iobs}{observed boundary contacts}

\item{phi}{numeric vector c(alpha, beta, gamma)}
}
\value{
sum of squared residuals
}
\description{
One-sided EM objective (Eq 18)
}
\examples{
t <- c(1, 2, 3)
lv_em_objective_one_sided(t, Sobs = c(0.2, 0.3, 0.5),
                          Iobs = c(0.1, 0.2, 0.4),
                          phi = c(1, 0.3, 0.5))
}
