% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lv_expression_palette}
\alias{lv_expression_palette}
\title{Default contact-intensity palette for embedding overlays}
\usage{
lv_expression_palette()
}
\value{
vector of 256 colours
}
\description{
Default contact-intensity palette for embedding overlays
}
\examples{
plot(1:10, col = lv_expression_palette()[seq(1, 256, length.out = 10)], pch = 19)
}
