% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp.R
\name{lv_project_two_sided}
\alias{lv_project_two_sided}
\title{Two-sided projection horizon (Eq 16)}
\arguments{
\item{Iobs}{observed I-boundary contacts}

\item{theta, gamma, alpha}{model parameters}

\item{delta_t}{extrapolation horizon}
}
\value{
numeric vector of projected increments
}
\description{
dI = (alpha/gamma - I) * (1 - exp(-(1-theta)\emph{gamma}delta_t)).
}
\examples{
lv_project_two_sided(Iobs = c(0.2, 0.8, 2), theta = 0.6, gamma = 0.3,
                     delta_t = 2, alpha = 1)
}
