% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lv_set_names}
\alias{lv_set_names}
\title{Set a vector's names to its own values}
\usage{
lv_set_names(x)
}
\arguments{
\item{x}{vector}
}
\value{
x with \code{names(x) <- x}
}
\description{
Set a vector's names to its own values
}
\examples{
lv_set_names(c("a", "b"))
}
