% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lv_val2col}
\alias{lv_val2col}
\title{Map numeric values onto a colour gradient}
\usage{
lv_val2col(
  x,
  gradientPalette = NULL,
  zlim = NULL,
  gradient.range.quantile = 0.95
)
}
\arguments{
\item{x}{numeric vector; nonnegative values use a sequential ramp,
signed values a diverging blue-grey-red ramp}

\item{gradientPalette}{optional explicit gradient}

\item{zlim}{optional value range}

\item{gradient.range.quantile}{quantile clip for automatic zlim}
}
\value{
colour vector (named if x was)
}
\description{
Map numeric values onto a colour gradient
}
\examples{
lv_val2col(c(-2, -1, 0, 1, 2))
}
