% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_hic_juicer.R
\name{read_hic}
\alias{read_hic}
\title{Read contacts from a Juicer .hic file}
\usage{
read_hic(file, resolution, chroms = NULL)
}
\arguments{
\item{file}{path to a .hic file}

\item{resolution}{resolution (bp) at which to pull contacts}

\item{chroms}{chromosomes to extract, e.g. \code{c("chr1", "chr2")};
cis-contacts per chromosome}
}
\value{
data.frame of contact pixels (chr1, x1, chr2, y1, count)
}
\description{
Read contacts from a Juicer .hic file
}
\examples{
\dontrun{
px <- read_hic("allpairs.hic", resolution = 5000, chroms = c("chr1", "chr2"))
}
}
