% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_hic.R
\name{read_mcool}
\alias{read_mcool}
\title{Read one resolution layer of a multi-resolution cooler file}
\usage{
read_mcool(file, resolution)
}
\arguments{
\item{file}{path to a .mcool file}

\item{resolution}{resolution (bp) selecting the group under
\code{/resolutions}}
}
\value{
\code{lv_contacts}
}
\description{
Read one resolution layer of a multi-resolution cooler file
}
\examples{
if (requireNamespace("hdf5r")) {
  f <- tempfile(fileext = ".mcool")
  h <- hdf5r::H5File$new(f, mode = "w")
  g <- h$create_group("resolutions")$create_group("10000")
  gc <- g$create_group("chroms"); gc[["name"]] <- "chr1"; gc[["length"]] <- 5e4
  gb <- g$create_group("bins")
  gb[["chrom"]] <- rep("chr1", 5)
  gb[["start"]] <- seq(0, 4e4, by = 1e4); gb[["end"]] <- seq(1e4, 5e4, by = 1e4)
  gp <- g$create_group("pixels")
  gp[["bin1_id"]] <- 0:1; gp[["bin2_id"]] <- 1:2; gp[["count"]] <- c(2, 1)
  h$close_all()
  x <- read_mcool(f, resolution = 10000)
  head(x$pixels)
}
}
